+++ srcs/manage_scene/rgb/manage_rgb.c
@@ -6,7 +6,7 @@
 /*   By: njooris <njooris@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/10 09:14:09 by njooris           #+#    #+#             */
-/*   Updated: 2025/10/28 13:52:58 by njooris          ###   ########.fr       */
+/*   Updated: 2025/10/29 16:34:52 by njooris          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,7 +91,6 @@ t_rgb	color_at(t_world w, t_ray r, uint32_t nb_bounce)
 	}
 	prepare_computations(h, r);
 	direct_color = shade_hit(w, *h);
-	//indirect_light = indirect_light_maker(h, w, NB_BOUNCE);
 	indirect_light = set_rgb(0, 0, 0);
 	free(inters.inters);
 	return (rgb_addition(direct_color, indirect_light));
