+++ srcs/parsing/parse_objs.c
@@ -6,7 +6,7 @@
 /*   By: njooris <njooris@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/15 11:14:01 by njooris           #+#    #+#             */
-/*   Updated: 2025/10/29 13:42:32 by njooris          ###   ########.fr       */
+/*   Updated: 2025/10/29 16:57:52 by njooris          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,6 +70,23 @@ int	free_line(char *line)
 	return (1);
 }
 
+int	build_objs(char *line_obj, t_world *world, int *count_obj)
+{
+	if (!ft_strncmp(line_obj, "sp", 2))
+	{
+		if (make_sp(&world->obj[*count_obj], line_obj))
+			return (free_line(line_obj));
+		*count_obj += 1;
+	}
+	if (!ft_strncmp(line_obj, "pl", 2))
+	{
+		if (make_pl(&world->obj[*count_obj], line_obj))
+			return (free_line(line_obj));
+		*count_obj += 1;
+	}
+	return (0);
+}
+
 int	make_objs(t_world *world, t_camera *c, int fd)
 {
 	int			count_obj;
@@ -83,11 +100,8 @@ int	make_objs(t_world *world, t_camera *c, int fd)
 	count_light = 0;
 	while (line_obj)
 	{
-		if (!ft_strncmp(line_obj, "sp", 2)
-			&& make_sp(&world->obj[count_obj++], line_obj))
-			return (free_line(line_obj));
-		if (!ft_strncmp(line_obj, "pl", 2) && make_pl(&world->obj[count_obj++], line_obj))
-			return (free_line(line_obj));
+		if (build_objs(line_obj, world, &count_obj))
+			return (1);
 		if (line_obj[0] == 'L'
 			&& make_light(&world->light[count_light++], line_obj))
 			return (free_line(line_obj));
