+++ srcs/parsing/parse_sphere.c
@@ -6,7 +6,7 @@
 /*   By: njooris <njooris@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/15 11:25:19 by njooris           #+#    #+#             */
-/*   Updated: 2025/10/29 13:43:18 by njooris          ###   ########.fr       */
+/*   Updated: 2025/10/29 16:38:10 by njooris          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ int	make_sp(t_obj *o, char *str)
 	*o = shape(SPHERE);
 	if (size_of_split(data) != 4 || get_coord(data[1], &coor)
 		|| check_is_digit_str(data[2]))
-	{	
+	{
 		ft_free_split(data);
 		return (1);
 	}
